requirements.txt
openai==1.57.4
typer-slim[shellingham]==0.15.4
jinja2==3.1.4
json5==0.10.0
pyyaml==6.0.2
markdown==3.7
gradio==5.9.1
retry==0.9.2